Fixing update bugs on multiple instances
