+++ custom_components/energidataservice/sensor.py
@@ -454,7 +454,7 @@ class EnergidataserviceSensor(SensorEntity):
         await super().async_added_to_hass()
         _LOGGER.debug("Added sensor '%s'", self._entity_id)
         await self.validate_data()
-        async_dispatcher_connect(self._hass, UPDATE_EDS, self.validate_data)
+        async_dispatcher_connect(self._hass, UPDATE_EDS.format(self._entry_id), self.validate_data)
 
     @property
     def unique_id(self):
